requirements.txt
torch
transformers
faiss-cpu
pdfplumber
nltk
torchvision
